// src/types/relatedCourses.ts
export type relatedCoursesType = {
  id?: number;
  title?: string;
  category?: string;
  categoryColor?: string;
  categoryBgColor?: string;
  price?: number;
  thumbnail?: string;
  review?: number;
  registeredStudents?: number;
  courseOwner?: string;
  level?: string;
  duration?: number;
  description?: string;
  name?: string;
  courseCount?: number;
};